+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix panic when processing accept message while peer is disconnected.
+- Removed stop-gap from receiving payments with open position.
+- Reduced min amount of 50k sats on receiving payments.
 
 ## [1.0.20] - 2023-06-16